# docs/api/API_swagger.yaml
definitions:
  github_com_tbxark_go-base-api_pkg_cdn_cdnmodels.UploadToken:
    properties:
      key:
        type: string
      token:
        type: string
      url:
        type: string
    type: object
  github_com_tbxark_go-base-api_pkg_dao_ent.User:
    properties:
      avatar:
        description: 头像
        type: string
      created_at:
        description: 创建时间
        type: integer
      flags:
        description: 标记位
        type: integer
      id:
        description: ID of the ent.
        type: integer
      phone:
        description: 手机号
        type: string
      remark:
        description: 备注
        type: string
      updated_at:
        description: 更新时间
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  github_com_tbxark_go-base-api_pkg_web.DataResponse-github_com_tbxark_go-base-api_pkg_cdn_cdnmodels_UploadToken:
    properties:
      data:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_cdn_cdnmodels.UploadToken'
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_AuthResponse:
    properties:
      data:
        $ref: '#/definitions/internal_biz_api.AuthResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_UpdateUserInfoResponse:
    properties:
      data:
        $ref: '#/definitions/internal_biz_api.UpdateUserInfoResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_UserInfoMeResponse:
    properties:
      data:
        $ref: '#/definitions/internal_biz_api.UserInfoMeResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_tbxark_go-base-api_pkg_web_webmodels.MessageResponse:
    properties:
      message:
        type: string
    type: object
  internal_biz_api.AuthResponse:
    properties:
      isNew:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_dao_ent.User'
    type: object
  internal_biz_api.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_web_webmodels.MessageResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_biz_api.UpdateUserInfoRequest:
    properties:
      avatar:
        type: string
      username:
        type: string
    type: object
  internal_biz_api.UpdateUserInfoResponse:
    properties:
      info:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_dao_ent.User'
    type: object
  internal_biz_api.UserInfoMeResponse:
    properties:
      info:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_dao_ent.User'
      inviter:
        $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_dao_ent.User'
    type: object
  internal_biz_api.WxMiniAuthRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  internal_biz_api.WxMiniBindPhoneRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
info:
  contact: {}
paths:
  /api/status:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_biz_api.MessageResponse'
      summary: 获取系统状态
      tags:
      - api
  /api/upload/token:
    get:
      consumes:
      - application/json
      parameters:
      - description: 文件名
        in: query
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_web.DataResponse-github_com_tbxark_go-base-api_pkg_cdn_cdnmodels_UploadToken'
      security:
      - ApiKeyAuth: []
      summary: 获取上传凭证
      tags:
      - api
  /api/user/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_UserInfoMeResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - api
  /api/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_biz_api.UpdateUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_UpdateUserInfoResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - api
  /api/wx/mini/auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_biz_api.WxMiniAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tbxark_go-base-api_pkg_web.DataResponse-internal_biz_api_AuthResponse'
      summary: 微信小程序登录
      tags:
      - api
  /api/wx/mini/bind/phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: 绑定信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_biz_api.WxMiniBindPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_biz_api.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: 绑定手机号
      tags:
      - api
securityDefinitions:
  ApiKeyAuth:
    description: JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"